Injected texturemanager - next would be to create an animationmodule and then a percentage rect

+++ PumpItUpEngine/PumpItUpEngine/ObjectList.cpp
@@ -9,7 +9,7 @@ const std::vector<Object>& ObjectList::GetObjectList() {
 
 int ObjectList::AddObject(Object object) {
 	std::lock_guard<std::mutex> guard(listMutex); // Lock the mutex
-	int objectId = objectList.size();
+	float objectId = objectList.size();
 	objectList.push_back(object);
 	return objectId;
 }
